+++ test/smokes/javasee/expectations.rb
@@ -1,6 +1,6 @@
 s = Runners::Testing::Smoke
 
-default_version = "0.1.3"
+default_version = "0.2.0"
 
 s.add_test(
   "success",