+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 [Full diff](https://github.com/sider/runners/compare/0.36.1...HEAD)
 
+- Bump Ruby from 2.7.1 to 2.7.2 [#1538](https://github.com/sider/runners/pull/1538)
+
 ## 0.36.1
 
 [Full diff](https://github.com/sider/runners/compare/0.36.0...0.36.1)